requirements.txt
certifi==2022.6.15
pulsar-client==2.10.1
pygame==2.1.2
six==1.16.0
